Xcode 8 updates

+++ Chinese flashcards/WordPack.m
@@ -95,7 +95,7 @@
     NSUInteger count = [_words count];
     for (NSUInteger i = 0; i < count; ++i) {
         NSInteger remainingCount = count - i;
-        NSInteger exchangeIndex = i + arc4random_uniform(remainingCount);
+        NSInteger exchangeIndex = i + arc4random_uniform((uint32_t)remainingCount);
         [_words exchangeObjectAtIndex:i withObjectAtIndex:exchangeIndex];
     }
 }
